// src/app/main/finance/Shared/Dtos/book-register-dto.ts
export class BankRegister {
    documentDate: Date;
    voucherNumber: string;
    comments: string;
    serviceEquipmentId: number;
    serviceEquipmentName: string;
    projectId: number;
    projectName: string;
    costCenterId: number;
    costCenterName: string;
    chequeNoStart: string;
    serialStart: string;
    noOfCheques: number;
}